Removing calls to Puppet.features.rubygems?
This makes the Gem usable in the latest Puppet versions.

The handling of RubyGems got revised in Puppet 3.0.1-rc1:
  http://projects.puppetlabs.com/issues/16757

The new policy is that either bundler and/or rubygems are guaranteed to
be loaded and initialized when the Puppet manifest is evaluated, making
it unnecessary for Puppet modules to load rubygems.

This new policy broke the puppet-trocla module. This is because
'Puppet.features.rubygems?' always evaluates to false now, which causes
the module to abort the manifest compilation with a message informing
about the necessity of RubyGems to be present.