keep trocla object around - addresses #18
Per puppet function call we now created a new Trocla object. This
is a) a very naive approach and b) obviously can lead to a lot of
inefficiency as we for example need to build up trocla each time
again. Also this means that we are running into problems like
opening a connection to a database system each time a trocla
lookup is done (and we never close the connection :-/).

The proper way to solve this is to make sure we don't create too
many trocla objects. With this change, we should now create once
a global trocla object PER puppet(-master/-server) process and
keep it around for the life time of such a process.