Setup github CI
This will make CI run sytax and lint validation on ruby and puppet code
and will run rspec tests whenever a push occurs or on pull requests.

This exposes some lint issues both for puppet and ruby, and errors out
on the spec tests. Fixing those will be done on subsequent commits.

The module has a travis-ci configuration file but CI there is not
working anymore since they stopped supporting free projects, so we're
replacing it entirely.